Integrate with Convenient Chests
This resolves the problem of the categories not being saved for chests
added by MegaStorage. The cause of the problem is that Convenient Chests
has an internal mapping of chest -> data. While saving, all MegaStorage
chests are temporarily replaced with regular chests, which do not have
the same data mapped to them.

This change makes use of a new API in Convenient Chests, which allows us
to copy this data to and from these temporary chests.